monitor - use built-in Angular control flow

+++ frontend/src/app/monitor/group/delete/monitor-group-delete-page.component.ts
@@ -1,4 +1,4 @@
-import { NgIf } from '@angular/common';
+import { inject } from '@angular/core';
 import { ChangeDetectionStrategy } from '@angular/core';
 import { Component } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
@@ -14,50 +14,54 @@ import { MonitorGroupDeletePageService } from './monitor-group-delete-page.servi
   selector: 'kpn-monitor-group-delete-page',
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: `
-    <kpn-page *ngIf="service.state() as state">
+    <kpn-page>
       <kpn-monitor-group-breadcrumb />
 
       <h1 i18n="@@monitor.group.delete.title">Monitor - delete group</h1>
 
-      <div *ngIf="state.response as response" class="kpn-form">
-        <div *ngIf="!response.result">
-          <p i18n="@@monitor.group.delete.group-not-found">Group not found</p>
-        </div>
-        <div *ngIf="response.result as page" class="kpn-form">
-          <p>
-            <span class="kpn-label" i18n="@@monitor.group.delete.name">
-              Name
-            </span>
-            {{ page.groupName }}
-          </p>
+      @if (service.state(); as state) {
+        @if (state.response; as response) {
+          @if (!response.result) {
+            <p class="kpn-error" i18n="@@monitor.group.delete.group-not-found">Group not found</p>
+          }
+          @if (response.result; as page) {
+            <div class="kpn-form">
+              <p>
+                <span class="kpn-label" i18n="@@monitor.group.delete.name"> Name </span>
+                {{ page.groupName }}
+              </p>
 
-          <p>
-            <span class="kpn-label" i18n="@@monitor.group.delete.description">
-              Description
-            </span>
-            {{ page.groupDescription }}
-          </p>
+              <p>
+                <span class="kpn-label" i18n="@@monitor.group.delete.description">
+                  Description
+                </span>
+                {{ page.groupDescription }}
+              </p>
 
-          <div *ngIf="page.routes.length as routeCount">
-            <div *ngIf="routeCount > 0" class="kpn-line">
-              <mat-icon svgIcon="warning" />
-              <span i18n="@@monitor.group.delete.warning">
-                The information of all routes ({{ routeCount }} route(s)) in the
-                group will also be deleted!
-              </span>
+              @if (page.routes.length; as routeCount) {
+                @if (routeCount > 0) {
+                  <div class="kpn-line">
+                    <mat-icon svgIcon="warning" />
+                    <span i18n="@@monitor.group.delete.warning">
+                      The information of all routes ({{ routeCount }} route(s)) in the group will
+                      also be deleted!
+                    </span>
+                  </div>
+                }
+              }
+
+              <div class="kpn-form-buttons">
+                <button mat-stroked-button (click)="service.delete(page.groupId)">
+                  <span class="kpn-warning" i18n="@@monitor.group.delete.action">
+                    Delete group
+                  </span>
+                </button>
+                <a routerLink="/monitor" i18n="@@action.cancel">Cancel</a>
+              </div>
             </div>
-          </div>
-
-          <div class="kpn-form-buttons">
-            <button mat-stroked-button (click)="service.delete(page.groupId)">
-              <span class="kpn-warning" i18n="@@monitor.group.delete.action">
-                Delete group
-              </span>
-            </button>
-            <a routerLink="/monitor" i18n="@@action.cancel">Cancel</a>
-          </div>
-        </div>
-      </div>
+          }
+        }
+      }
       <kpn-sidebar sidebar />
     </kpn-page>
   `,
@@ -67,12 +71,11 @@ import { MonitorGroupDeletePageService } from './monitor-group-delete-page.servi
     MatButtonModule,
     MatIconModule,
     MonitorGroupBreadcrumbComponent,
-    NgIf,
     PageComponent,
     RouterLink,
     SidebarComponent,
   ],
 })
 export class MonitorGroupDeletePageComponent {
-  constructor(protected service: MonitorGroupDeletePageService) {}
+  protected readonly service = inject(MonitorGroupDeletePageService);
 }
